LayoutTests/fast/canvas/html-in-canvas/drawElementImage-2arg-returns-identity-matrix.html

<!DOCTYPE html><!-- webkit-test-runner [ CanvasDrawElementEnabled=true ] -->
<title>drawElementImage returns identity matrix in TB1b (TB3a will return real matrix)</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<canvas id="c" width="100" height="100" layoutsubtree>
  <div id="child" style="background:green; width:30px; height:30px;"></div>
</canvas>
<script>
function nextPaint() {
    return new Promise(r => requestAnimationFrame(() => requestAnimationFrame(r)));
}

promise_test(async () => {
    const canvas = document.getElementById("c");
    const child = document.getElementById("child");
    await nextPaint();
    const m = canvas.getContext("2d").drawElementImage(child, 0, 0);
    assert_true(m instanceof DOMMatrix, "returns a DOMMatrix");
    assert_true(m.is2D, "matrix is 2D");
    assert_true(m.isIdentity, "TB1b returns identity matrix; TB3a will return the real alignment matrix");
}, "drawElementImage returns an identity DOMMatrix in TB1b");
</script>
